+++ packages/ui-kit/src/components/PieChart/PieChart.tsx
@@ -60,12 +60,16 @@ export const PieChartComponent = React.forwardRef<HTMLDivElement, PieChartProps>
       data: leaderboardData,
       isLoading: leaderboardIsLoading,
       error: leaderboardHasError
-    } = useLeaderboard({ ...query })
+    } = useLeaderboard({ ...query, dimensions: [query?.dimension ?? { columnName: '' }], enabled: !isStatic })
 
     /**
      * Fetches the counter data from the API
      */
-    const { data: counterData, isLoading: counterIsLoading, error: counterHasError } = useCounter({ ...query })
+    const {
+      data: counterData,
+      isLoading: counterIsLoading,
+      error: counterHasError
+    } = useCounter({ ...query, enabled: !isStatic })
 
     const isLoading = leaderboardIsLoading || counterIsLoading
 
@@ -258,7 +262,7 @@ export const PieChartComponent = React.forwardRef<HTMLDivElement, PieChartProps>
           (hasError?.name === 'AccessTokenError' ||
             !query.metric ||
             !query.timeRange ||
-            !query.dimensions ||
+            !query.dimension ||
             !query.rowLimit)
         ) {
           // console.error(