+++ packages/ui-kit/src/hooks/useTimeSeries/useTimeSeries.ts
@@ -18,6 +18,7 @@ export const useTimeSeries = (props: TimeSeriesQueryProps): UseQueryProps<TimeSe
     granularity,
     filters: filtersFromProp,
     refetchInterval,
+    enabled: enabledProp = true,
     retry,
     timeZone
   } = props
@@ -38,7 +39,7 @@ export const useTimeSeries = (props: TimeSeriesQueryProps): UseQueryProps<TimeSe
 
   const filters = filtersFromProp ?? filtersFromProvider
 
-  const enabled = accessToken != null
+  const enabled = accessToken != null && enabledProp
 
   // Log error if no access token provided and metric is provided
   if (!enabled && metric) {