+++ packages/ui-kit/src/hooks/useTopValues/useTopValues.ts
@@ -20,6 +20,7 @@ export const useTopValues = ({
   maxValues,
   refetchInterval,
   retry,
+  enabled: enabledProp = true,
   timeZone
 }: TopValuesQueryProps): UseQueryProps<TopValuesQuery> => {
   const log = useLog()
@@ -34,10 +35,10 @@ export const useTopValues = ({
   // Get access token first from props, then if it is not provided via prop get it from provider
   const accessToken = accessTokenFromProp ?? accessTokenFromProvider
 
-  const enabled = accessToken != null
+  const enabled = accessToken != null && enabledProp
 
   // Log error if no access token provided and metric is provided
-  if (!enabled && metric) {
+  if (accessToken == null && metric) {
     log.error(accessTokenError ?? 'No access token provided.')
   }