+++ app/examples/react-17/CHANGELOG.md
@@ -1,5 +1,33 @@
 # react-17
 
+## 0.1.4
+
+### Patch Changes
+
+- 9797ad6: Refactored PieChart Dimension query prop on receive a single dimension
+- Updated dependencies [30e01c3]
+- Updated dependencies [9797ad6]
+- Updated dependencies [1ff7296]
+- Updated dependencies [33eca9e]
+  - @propeldata/ui-kit@0.5.0
+
+## 0.1.3
+
+### Patch Changes
+
+- Updated dependencies [76537de]
+  - @propeldata/ui-kit@0.4.1
+
+## 0.1.2
+
+### Patch Changes
+
+- Updated dependencies [f8041e7]
+- Updated dependencies [84a9f9b]
+- Updated dependencies [70ad890]
+- Updated dependencies [d12f0e1]
+  - @propeldata/ui-kit@0.4.0
+
 ## 0.1.1
 
 ### Patch Changes