+++ packages/ui-kit/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @propeldata/ui-kit
 
+## 0.5.0
+
+### Minor Changes
+
+- 33eca9e: Add SQL query hook
+
+### Patch Changes
+
+- 30e01c3: Fixed font-family not applying to piechart
+- 9797ad6: Refactored PieChart Dimension query prop on receive a single dimension
+- 1ff7296: Fixed component and type exports
+
+## 0.4.1
+
+### Patch Changes
+
+- 76537de: Export useDataGrid and useRecordsById
+
+## 0.4.0
+
+### Minor Changes
+
+- f8041e7: Fixed static mode fetching data bug
+- 84a9f9b: Add Records By Id query hook
+- d12f0e1: Added SimpleFilter component and FilterProvider
+
+### Patch Changes
+
+- 70ad890: Added LIKE and NOT_LIKE filter operators
+
 ## 0.3.1
 
 ### Patch Changes